Add LocationMessageCell
